requirements.txt
streamlit
pandas
dnspython
requests
email-validator
tqdm
asyncio
